添加浏览量统计

+++ article/views.py
@@ -1,4 +1,5 @@
 import markdown
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.core.paginator import Paginator
 from django.http import HttpResponse
@@ -26,6 +27,10 @@ def article_list(request):
 
 def article_detail(request, article_id):
     article = ArticlePost.objects.get(id=article_id)
+    # 浏览量 +1
+    if request.user != article.author:
+        article.total_views += 1
+        article.save(update_fields=['total_views'])
     article.body = markdown.markdown(article.body, extensions=[
         "markdown.extensions.extra",
         "markdown.extensions.codehilite",
@@ -34,6 +39,7 @@ def article_detail(request, article_id):
     return render(request, "article/detail.html", context)
 
 
+@login_required(login_url='/userprofile/login/')
 def article_create(request):
     # user = User.objects.get(id=article_id)
     # 判断用户是否提交数据
@@ -56,23 +62,33 @@ def article_create(request):
         return render(request, 'article/create.html', context)
 
 
+@login_required(login_url="/userprofile/login/")
 def article_delete(request, article_id):
     article = ArticlePost.objects.get(id=article_id)
+    if request.user != article.author:
+        return HttpResponse("抱歉，你无权修改这篇文章。")
     article.delete()
     return redirect("article:article_list")
 
 
+@login_required(login_url='/userprofile/login/')
 def article_safe_delete(request, article_id):
     if request.method == 'POST':
         article = ArticlePost.objects.get(id=article_id)
+        if request.user != article.author:
+            return HttpResponse("抱歉，你无删除这篇文章。")
         article.delete()
         return redirect("article:article_list")
     else:
         return HttpResponse("仅允许post请求")
 
 
+@login_required(login_url='/userprofile/login/')
 def article_update(request, article_id):
     article = ArticlePost.objects.get(id=article_id)
+    # 过滤非作者的用户
+    if request.user != article.author:
+        return HttpResponse("抱歉，你无权修改这篇文章。")
     if request.method == "POST":
         article_post_form = ArticlePostForm(data=request.POST)
         if article_post_form.is_valid():
